requirements.txt
toolz
colour
svgwrite
cairosvg
Pillow
git+https://github.com/chalk-diagrams/planar
latextools
loguru
typing-extensions
